manicure_bot/requirements.txt
aiogram==3.24.0
python-dotenv==1.2.1
aiosqlite==0.22.1
APScheduler==3.11.2
